Moving more implementation pieces to the device side
