+++ src/DeckSurf.SDK/Models/Devices/StreamDeckMini2022.cs
@@ -78,6 +78,12 @@ namespace DeckSurf.SDK.Models.Devices
         }
 
         /// <inheritdoc/>
+        public override bool SetScreen(byte[] image, int offset, int width, int height)
+        {
+            return false;
+        }
+
+        /// <inheritdoc/>
         protected override ButtonPressEventArgs HandleKeyPress(IAsyncResult result, byte[] keyPressBuffer)
         {
             int bytesRead = this.UnderlyingInputStream.EndRead(result);
