+++ src/DeckSurf.SDK/Models/Devices/StreamDeckNeo.cs
@@ -5,6 +5,7 @@
 using System;
 using System.Drawing;
 using System.Drawing.Imaging;
+using System.Linq;
 using DeckSurf.SDK.Util;
 
 namespace DeckSurf.SDK.Models.Devices
@@ -77,6 +78,53 @@ namespace DeckSurf.SDK.Models.Devices
                 binaryIteration[0],
                 binaryIteration[1],
             ];
+        }
+
+        /// <inheritdoc/>
+        public override bool SetScreen(byte[] image, int offset, int width, int height)
+        {
+            byte[] binaryOffset = DataHelpers.GetLittleEndianBytesFromInt(offset);
+            byte[] binaryWidth = DataHelpers.GetLittleEndianBytesFromInt(width);
+            byte[] binaryHeight = DataHelpers.GetLittleEndianBytesFromInt(height);
+
+            var iteration = 0;
+            var remainingBytes = image.Length;
+
+            using var stream = this.Open();
+            while (remainingBytes > 0)
+            {
+                var sliceLength = Math.Min(remainingBytes, (this.PacketSize - this.ScreenImageHeaderSize));
+                var bytesSent = iteration * (this.PacketSize - this.ScreenImageHeaderSize);
+
+                byte isLastChunk = sliceLength == remainingBytes ? (byte)1 : (byte)0;
+
+                var binaryLength = DataHelpers.GetLittleEndianBytesFromInt(sliceLength);
+                var binaryIteration = DataHelpers.GetLittleEndianBytesFromInt(iteration);
+
+                byte[] header =
+                [
+                    0x02,
+                    0x0B,
+                    0x00,
+                    isLastChunk,
+                    binaryLength[0],
+                    binaryLength[1],
+                    binaryIteration[0],
+                    binaryIteration[1],
+                ];
+
+                var payload = header.Concat(new ArraySegment<byte>(image, bytesSent, sliceLength)).ToArray();
+                var padding = new byte[this.PacketSize - payload.Length];
+
+                var finalPayload = payload.Concat(padding).ToArray();
+
+                stream.Write(finalPayload);
+
+                remainingBytes -= sliceLength;
+                iteration++;
+            }
+
+            return true;
         }
 
         /// <inheritdoc/>
